code  to print the last digit of the number stored in the variable n.

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -31,16 +31,16 @@ int main(void)
 	n = rand() - RAND_MAX / 2;
 
 	/* your code goes there */
-	int last_digit= n%10;
+	long last_digit = n%10;
 
 	if (last_digit > 5)
-		printf("Last digit of %lu is %lu and is greater than 5\n"n,last_digit);
+		printf("Last digit of %li is %li and is greater than 5\n",n,last_digit);
 
 	else if (last_digit=0)
-		printf("Last digit of %lu is %lu and is 0\n",n,last_digit);
+		printf("Last digit of %li is %li and is 0\n",n,last_digit);
 
 	else if ((last_digit < 6) && != 0)
-		printf("Last digit of %lu is %lu and is less than 6 and not 0\n",n,last_digit);
+		printf("Last digit of %li is %li and is less than 6 and not 0\n",n,last_digit);
 
 	return (0);
